Update how the built image is run in GitHub Actions
On Ubuntu 22.04 GitHub Actions runners cgroups v2 are in use and
changes are necessary to successfully start this systemd enabled image.
The default namespace for cgroups v2 is `private` while the default for
cgroups v1 was `host` so we must manually change the cgroups namespace
to have the same behavior. We must also change the mounted cgroup
volume from read-only to read-write permissions.